refactor(workflows): split schemas.rs (1429 lines) into schemas/ directory
Splits the monolithic schemas.rs into five focused sub-modules under
src/openhuman/workflows/schemas/:

- wire_types.rs   (~200 lines) — RPC param/result structs, WorkflowSummary
- helpers.rs      (~80 lines)  — config/workspace resolution, deserialize_params/to_json
- run_machinery.rs (~230 lines) — spawn_workflow_run_background, await_run_outcome
- handlers.rs     (~240 lines) — thin handle_* dispatcher functions
- controller_schemas.rs (~300 lines) — workflows_schemas match + all_* registry fns

External API (all_workflows_controller_schemas, all_workflows_registered_controllers,
workflows_schemas, resolve_workspace_dir, spawn_workflow_run_background,
await_run_outcome) is fully preserved; existing callers require no changes.
