+++ src/openhuman/channels/runtime/dispatch/processor.rs
@@ -1,650 +1,41 @@
-//! Channel runtime loop and message processing.
+//! Core message processing loop for the channel runtime.
+//!
+//! Contains:
+//! * [`channel_has_approval_surface`] — gate controlling per-channel approval
+//!   context scoping.
+//! * [`try_route_approval_reply`] — intercepts yes/no approval replies before
+//!   dispatching a fresh agent turn.
+//! * [`process_channel_message`] — full per-message pipeline: typing, ACK
+//!   reaction, history, agent turn, draft updates, reply.
+//! * [`run_message_dispatch_loop`] — bounded-concurrency worker loop that feeds
+//!   messages into [`process_channel_message`].
 
 use crate::core::event_bus::{
     publish_global, request_native_global, DomainEvent, NativeRequestError,
 };
 use crate::openhuman::agent::bus::{AgentTurnRequest, AgentTurnResponse, AGENT_RUN_TURN_METHOD};
-use crate::openhuman::agent::harness::definition::{
-    AgentDefinition, AgentDefinitionRegistry, ToolScope,
-};
 use crate::openhuman::agent::progress::AgentProgress;
 use crate::openhuman::channels::context::{
     build_memory_context, compact_sender_history, conversation_history_key,
     conversation_memory_key, is_context_window_overflow_error, ChannelRuntimeContext,
-    CHANNEL_TYPING_REFRESH_INTERVAL_SECS, MAX_CHANNEL_HISTORY,
 };
 use crate::openhuman::channels::providers::telegram::TELEGRAM_APPROVAL_CLIENT_ID;
 use crate::openhuman::channels::routes::{
     get_or_create_provider, get_route_selection, handle_runtime_command_if_needed,
 };
 use crate::openhuman::channels::traits;
-use crate::openhuman::channels::{Channel, SendMessage};
-use crate::openhuman::composio::fetch_connected_integrations;
-use crate::openhuman::config::Config;
+use crate::openhuman::channels::SendMessage;
 use crate::openhuman::inference::provider::{self, ChatMessage};
-use crate::openhuman::tools::{orchestrator_tools, Tool};
 use crate::openhuman::util::truncate_with_ellipsis;
-use std::collections::HashSet;
 use std::sync::Arc;
 use std::time::{Duration, Instant};
 use tokio_util::sync::CancellationToken;
 
-/// Maximum characters shown in the debug reply println. Large enough to not truncate
-/// real responses while keeping terminal output readable.
-const REPLY_LOG_TRUNCATE_CHARS: usize = 200;
-
-/// Returns `true` if `s` contains any of the given substrings.
-#[inline]
-fn contains_any(s: &str, words: &[&str]) -> bool {
-    words.iter().any(|w| s.contains(w))
-}
-
-/// Returns `true` if `s` starts with any of the given prefixes.
-#[inline]
-fn starts_with_any(s: &str, prefixes: &[&str]) -> bool {
-    prefixes.iter().any(|p| s.starts_with(p))
-}
-
-/// Build the per-turn `[Channel context]` block prepended to the user
-/// message for non-web inbound channels (e.g. Telegram, Discord, Slack).
-///
-/// Surfaces the active channel and reply target so the model knows
-/// where it is talking and can route any tool side-effects (notably
-/// `cron_add`) back to the same chat instead of defaulting to the
-/// in-app web stream. See issue #928.
-///
-/// Returns an empty string for web/cli turns (the desktop UI is the
-/// default delivery surface, no hint needed).
-fn build_channel_context_block(msg: &traits::ChannelMessage) -> String {
-    let channel = msg.channel.trim();
-    if channel.is_empty()
-        || channel.eq_ignore_ascii_case("web")
-        || channel.eq_ignore_ascii_case("cli")
-    {
-        return String::new();
-    }
-
-    let reply_target = msg.reply_target.trim();
-    if reply_target.is_empty() {
-        return String::new();
-    }
-
-    format!(
-        "[Channel context]\n\
-         You are responding via the \"{channel}\" channel. Reply target: \"{reply_target}\".\n\
-         For any cron/scheduled reminder you create with `cron_add`, set `delivery` to \
-         `{{ \"mode\": \"announce\", \"channel\": \"{channel}\", \"to\": \"{reply_target}\" }}` \
-         so the reminder is delivered back here instead of the in-app web stream. \
-         Only fall back to the default proactive delivery if the user explicitly asks for \
-         in-app/desktop notification.\n\n"
-    )
-}
-
-/// Pick a contextual acknowledgment emoji for an inbound message.
-///
-/// Intent categories are checked in priority order. Within each category two
-/// emoji options are defined; a cheap deterministic index (based on message
-/// length + first char value) selects between them so that similar messages
-/// don't always produce the identical reaction.
-///
-/// All emojis used here are in Telegram's standard (non-premium) reaction set.
-fn select_acknowledgment_reaction(content: &str) -> &'static str {
-    let l = content.to_lowercase();
-
-    // Deterministic variant (0 or 1) — avoids true randomness while giving variety.
-    let v = content
-        .len()
-        .wrapping_add(content.chars().next().map_or(0, |c| c as usize))
-        & 1;
-
-    let opts: &[&str] = if contains_any(&l, &["thank", "thx", "appreciate", "grateful", "cheers"]) {
-        // Gratitude
-        &["❤️", "🙏"]
-    } else if contains_any(
-        &l,
-        &[
-            "amazing",
-            "awesome",
-            "incredible",
-            "love it",
-            "congrat",
-            "!!",
-        ],
-    ) {
-        // Excitement / celebration
-        &["🔥", "🎉"]
-    } else if contains_any(
-        &l,
-        &[
-            "price", "btc", "eth", "crypto", "trade", "pump", "dump", "market", "token", "wallet",
-            "defi", "nft", "sol", "bnb",
-        ],
-    ) {
-        // Crypto / finance
-        &["💯", "⚡"]
-    } else if contains_any(
-        &l,
-        &[
-            "code",
-            "function",
-            "api",
-            "deploy",
-            "build",
-            "debug",
-            "script",
-            "git",
-            "rust",
-            "python",
-            "js",
-            "typescript",
-        ],
-    ) {
-        // Technical / dev
-        &["👨‍💻", "🤓"]
-    } else if starts_with_any(
-        &l,
-        &[
-            "hi",
-            "hello",
-            "hey",
-            "sup",
-            "good morning",
-            "good evening",
-            "good afternoon",
-        ],
-    ) || l == "yo"
-        || l.starts_with("yo ")
-    {
-        // Greeting
-        &["🤗", "😁"]
-    } else if l.contains('?')
-        || starts_with_any(
-            &l,
-            &[
-                "how",
-                "what",
-                "why",
-                "when",
-                "where",
-                "who",
-                "can you",
-                "could you",
-                "would you",
-                "is ",
-                "are ",
-                "do you",
-                "does",
-            ],
-        )
-    {
-        // Question / help request
-        &["🤔", "✍️"]
-    } else {
-        // Default — "seen, on it"
-        &["👀", "✍️"]
-    };
-
-    opts[v % opts.len()]
-}
-
-fn log_worker_join_result(result: Result<(), tokio::task::JoinError>) {
-    if let Err(error) = result {
-        tracing::error!("Channel message worker crashed: {error}");
-    }
-}
-
-fn spawn_scoped_typing_task(
-    channel: Arc<dyn Channel>,
-    recipient: String,
-    cancellation_token: CancellationToken,
-) -> tokio::task::JoinHandle<()> {
-    let stop_signal = cancellation_token;
-    let refresh_interval = Duration::from_secs(CHANNEL_TYPING_REFRESH_INTERVAL_SECS);
-    let handle = tokio::spawn(async move {
-        loop {
-            tokio::select! {
-                () = stop_signal.cancelled() => break,
-                _ = tokio::time::sleep(refresh_interval) => {
-                    if let Err(e) = channel.start_typing(&recipient).await {
-                        tracing::debug!("Failed to start typing on {}: {e}", channel.name());
-                    }
-                }
-            }
-        }
-
-        if let Err(e) = channel.stop_typing(&recipient).await {
-            tracing::debug!("Failed to stop typing on {}: {e}", channel.name());
-        }
-    });
-
-    handle
-}
-
-/// Per-turn scoping fields derived from the active agent definition.
-///
-/// Carries the three new fields that get spliced into [`AgentTurnRequest`]
-/// in [`process_channel_message`]. Constructed by [`resolve_target_agent`]
-/// after reading `config.onboarding_completed`, looking up the matching
-/// definition in [`AgentDefinitionRegistry`], and synthesising any
-/// per-turn delegation tools the agent needs.
-struct AgentScoping {
-    target_agent_id: Option<String>,
-    visible_tool_names: Option<HashSet<String>>,
-    extra_tools: Vec<Box<dyn Tool>>,
-}
-
-impl AgentScoping {
-    /// Empty scoping — preserves the legacy "every tool in the global
-    /// registry is visible" behaviour. Returned when the registry isn't
-    /// initialised yet (early startup) or when the target agent
-    /// definition isn't found, so the channel layer never crashes the
-    /// runtime over a routing miss.
-    fn unscoped() -> Self {
-        Self {
-            target_agent_id: None,
-            visible_tool_names: None,
-            extra_tools: Vec::new(),
-        }
-    }
-}
-
-/// Decide which agent should run for this channel turn and build the
-/// matching tool-scoping payload.
-///
-/// All channel turns route directly to the `orchestrator` agent. The
-/// welcome agent has been removed; the Joyride walkthrough in the
-/// frontend handles onboarding UI instead.
-///
-/// On any failure path (missing registry, missing definition, missing
-/// orchestrator delegation targets) the function logs and returns
-/// [`AgentScoping::unscoped`], which lets the turn run with the legacy
-/// unfiltered behaviour rather than failing the whole message.
-async fn resolve_target_agent(channel: &str) -> AgentScoping {
-    let config = match Config::load_or_init().await {
-        Ok(c) => c,
-        Err(err) => {
-            tracing::warn!(
-                channel = %channel,
-                error = %err,
-                "[dispatch::routing] failed to load config — falling back to unscoped turn"
-            );
-            return AgentScoping::unscoped();
-        }
-    };
-
-    let target_id = "orchestrator";
-
-    tracing::info!(
-        channel = %channel,
-        target_agent = target_id,
-        ui_onboarding_completed = config.onboarding_completed,
-        "[dispatch::routing] selected target agent"
-    );
-
-    let registry = match AgentDefinitionRegistry::global() {
-        Some(reg) => reg,
-        None => {
-            tracing::warn!(
-                channel = %channel,
-                target_agent = target_id,
-                "[dispatch::routing] AgentDefinitionRegistry not initialised — falling back to unscoped turn"
-            );
-            return AgentScoping::unscoped();
-        }
-    };
-
-    let definition = match registry.get(target_id) {
-        Some(def) => def,
-        None => {
-            tracing::warn!(
-                channel = %channel,
-                target_agent = target_id,
-                "[dispatch::routing] target agent not in registry — falling back to unscoped turn"
-            );
-            return AgentScoping::unscoped();
-        }
-    };
-
-    // Synthesise per-turn delegation tools when the target agent has a
-    // `subagents = [...]` field. Today only the orchestrator does, but
-    // the helper is agent-agnostic so future agents that delegate
-    // (e.g. a custom workspace-override planner that subdivides work)
-    // pick this up for free.
-    //
-    // Wrap the Composio fetch in a 3-second timeout so a slow/unresponsive
-    // Composio API can never block turn dispatch indefinitely.
-    const COMPOSIO_FETCH_TIMEOUT_SECS: u64 = 3;
-    let extra_tools = if !definition.subagents.is_empty() {
-        let connected = match tokio::time::timeout(
-            Duration::from_secs(COMPOSIO_FETCH_TIMEOUT_SECS),
-            fetch_connected_integrations(&config),
-        )
-        .await
-        {
-            Ok(list) => list,
-            Err(_) => {
-                tracing::warn!(
-                    channel = %channel,
-                    target_agent = target_id,
-                    "[dispatch::routing] Composio fetch timed out after {}s — proceeding without connected integrations",
-                    COMPOSIO_FETCH_TIMEOUT_SECS
-                );
-                Vec::new()
-            }
-        };
-        tracing::debug!(
-            channel = %channel,
-            target_agent = target_id,
-            connected_integration_count = connected.len(),
-            "[dispatch::routing] fetched connected integrations for delegation expansion"
-        );
-        orchestrator_tools::collect_orchestrator_tools(definition, registry, &connected)
-    } else {
-        Vec::new()
-    };
-
-    let visible_tool_names = build_visible_tool_set(definition, &extra_tools);
-
-    tracing::debug!(
-        channel = %channel,
-        target_agent = target_id,
-        named_tool_count = match &definition.tools {
-            ToolScope::Named(names) => names.len(),
-            ToolScope::Wildcard => 0,
-        },
-        extra_tool_count = extra_tools.len(),
-        visible_tool_count = visible_tool_names.as_ref().map(|s| s.len()).unwrap_or(0),
-        "[dispatch::routing] assembled tool scoping for turn"
-    );
-
-    AgentScoping {
-        target_agent_id: Some(target_id.to_string()),
-        visible_tool_names,
-        extra_tools,
-    }
-}
-
-/// Build the visible-tool whitelist for an agent.
-///
-/// The set is the union of:
-/// * every tool name in the agent's `[tools] named = [...]` list
-///   (when the scope is [`ToolScope::Named`]); and
-/// * every name produced by the per-turn synthesised delegation tools
-///   in `extra_tools` (e.g. `research`, `plan`,
-///   `delegate_to_integrations_agent`).
-///
-/// When the agent's tool scope is [`ToolScope::Wildcard`] **and** there
-/// are no `extra_tools`, returns `None` to preserve the legacy
-/// "everything visible" semantics — a `Wildcard` agent that delegates
-/// nothing should still see the full registry. When `Wildcard` is
-/// combined with non-empty extras (an unusual but legal combination),
-/// the legacy unfiltered behaviour also wins because the wildcard
-/// implicitly covers anything in the registry plus the extras.
-fn build_visible_tool_set(
-    definition: &AgentDefinition,
-    extra_tools: &[Box<dyn Tool>],
-) -> Option<HashSet<String>> {
-    match &definition.tools {
-        ToolScope::Wildcard => None,
-        ToolScope::Named(names) => {
-            let mut set: HashSet<String> = names.iter().cloned().collect();
-            for tool in extra_tools {
-                set.insert(tool.name().to_string());
-            }
-            Some(set)
-        }
-    }
-}
-
-#[cfg(test)]
-mod scoping_tests {
-    //! Pure-function unit tests for the agent-scoping helpers added by
-    //! the #525/#526 fix. These exercise the synchronous logic without
-    //! touching the real `Config::load_or_init` disk read or the global
-    //! `AgentDefinitionRegistry`, so they can run in any environment.
-    //!
-    //! End-to-end exercise of the dispatch path is covered by the
-    //! existing `runtime_dispatch::dispatch_routes_through_agent_run_turn_
-    //! bus_handler` integration test, which still passes after the new
-    //! fields landed (the resolver gracefully falls back to
-    //! `AgentScoping::unscoped()` when no orchestrator is registered in
-    //! the test environment).
-
-    use super::*;
-    use crate::openhuman::agent::harness::definition::{
-        DefinitionSource, ModelSpec, PromptSource, SandboxMode,
-    };
-    use crate::openhuman::tools::traits::{PermissionLevel, Tool, ToolCategory, ToolResult};
-    use async_trait::async_trait;
-
-    /// Minimal owned tool stub — just enough for `build_visible_tool_set`
-    /// to read its `name()`.
-    struct StubTool {
-        name: &'static str,
-    }
-
-    #[async_trait]
-    impl Tool for StubTool {
-        fn name(&self) -> &str {
-            self.name
-        }
-        fn description(&self) -> &str {
-            "stub"
-        }
-        fn parameters_schema(&self) -> serde_json::Value {
-            serde_json::json!({"type": "object"})
-        }
-        fn category(&self) -> ToolCategory {
-            ToolCategory::System
-        }
-        fn permission_level(&self) -> PermissionLevel {
-            PermissionLevel::None
-        }
-        async fn execute(&self, _args: serde_json::Value) -> anyhow::Result<ToolResult> {
-            Ok(ToolResult::success("ok"))
-        }
-    }
-
-    fn def_with_scope(scope: ToolScope) -> AgentDefinition {
-        AgentDefinition {
-            id: "test_agent".into(),
-            when_to_use: "test".into(),
-            display_name: None,
-            system_prompt: PromptSource::Inline(String::new()),
-            omit_identity: true,
-            omit_memory_context: true,
-            omit_safety_preamble: true,
-            omit_skills_catalog: true,
-            omit_profile: true,
-            omit_memory_md: true,
-            model: ModelSpec::Inherit,
-            temperature: 0.4,
-            tools: scope,
-            disallowed_tools: vec![],
-            skill_filter: None,
-            extra_tools: vec![],
-            max_iterations: 8,
-            iteration_policy: Default::default(),
-            max_result_chars: None,
-            timeout_secs: None,
-            sandbox_mode: SandboxMode::None,
-            background: false,
-            subagents: vec![],
-            delegate_name: None,
-            agent_tier: crate::openhuman::agent::harness::definition::AgentTier::Worker,
-            source: DefinitionSource::Builtin,
-        }
-    }
-
-    /// `ToolScope::Wildcard` must yield `None` — the prompt builder
-    /// treats `None` as "no filter, every tool visible", which is the
-    /// correct behaviour for agents like `integrations_agent` that want the
-    /// full skill-category catalogue. Even when extras are present, a
-    /// wildcard agent should not start filtering.
-    #[test]
-    fn wildcard_scope_yields_none_filter() {
-        let def = def_with_scope(ToolScope::Wildcard);
-        let extras: Vec<Box<dyn Tool>> = vec![Box::new(StubTool { name: "research" })];
-        assert!(build_visible_tool_set(&def, &extras).is_none());
-        assert!(build_visible_tool_set(&def, &[]).is_none());
-    }
-
-    /// `ToolScope::Named` with no extras returns exactly the named set.
-    /// For agents with a narrow tool scope (e.g. 2 tools in TOML,
-    /// no delegation, no extras) → 2 entries in the visibility whitelist.
-    #[test]
-    fn named_scope_without_extras_returns_named_only() {
-        let def = def_with_scope(ToolScope::Named(vec![
-            "memory_recall".into(),
-            "ask_user_clarification".into(),
-        ]));
-        let set = build_visible_tool_set(&def, &[]).expect("named scope yields Some");
-        assert_eq!(set.len(), 2);
-        assert!(set.contains("memory_recall"));
-        assert!(set.contains("ask_user_clarification"));
-    }
-
-    /// `ToolScope::Named` with extras returns the union of the TOML
-    /// named list and the extras' names. This is the orchestrator's
-    /// path: direct tools from the TOML + the synthesised delegation
-    /// tools (`research`, `plan`, `delegate_to_integrations_agent`)
-    /// → all of them visible to the orchestrator's LLM. The stub
-    /// names in this test are arbitrary; they exercise the union
-    /// logic, not the real synthesiser.
-    #[test]
-    fn named_scope_with_extras_returns_union() {
-        let def = def_with_scope(ToolScope::Named(vec![
-            "query_memory".into(),
-            "ask_user_clarification".into(),
-            "spawn_subagent".into(),
-        ]));
-        let extras: Vec<Box<dyn Tool>> = vec![
-            Box::new(StubTool { name: "research" }),
-            Box::new(StubTool {
-                name: "delegate_gmail",
-            }),
-            Box::new(StubTool {
-                name: "delegate_github",
-            }),
-        ];
-        let set = build_visible_tool_set(&def, &extras).expect("named scope yields Some");
-        assert_eq!(set.len(), 6);
-        assert!(set.contains("query_memory"));
-        assert!(set.contains("ask_user_clarification"));
-        assert!(set.contains("spawn_subagent"));
-        assert!(set.contains("research"));
-        assert!(set.contains("delegate_gmail"));
-        assert!(set.contains("delegate_github"));
-    }
-
-    /// Empty `Named` list with extras still yields `Some` containing
-    /// just the extras — useful for hypothetical agents that only
-    /// reach the world via delegation, with no direct tools.
-    #[test]
-    fn empty_named_with_extras_returns_extras_only() {
-        let def = def_with_scope(ToolScope::Named(vec![]));
-        let extras: Vec<Box<dyn Tool>> = vec![Box::new(StubTool {
-            name: "delegate_only",
-        })];
-        let set = build_visible_tool_set(&def, &extras).expect("named scope yields Some");
-        assert_eq!(set.len(), 1);
-        assert!(set.contains("delegate_only"));
-    }
-
-    /// Empty `Named` list with no extras yields an empty `Some(set)` —
-    /// effectively "no tools visible". The prompt loop's `is_visible`
-    /// helper treats `Some(empty)` differently from `None`: the former
-    /// means "filter active, nothing matches" so the LLM gets an empty
-    /// tool list, while the latter means "no filter at all".
-    #[test]
-    fn empty_named_with_no_extras_returns_empty_set() {
-        let def = def_with_scope(ToolScope::Named(vec![]));
-        let set = build_visible_tool_set(&def, &[]).expect("named scope yields Some");
-        assert!(set.is_empty());
-    }
-
-    /// Duplicate names across named + extras are de-duplicated by the
-    /// HashSet — no double-counting if a workspace override happens to
-    /// list a delegation tool name in the direct `named` list too.
-    #[test]
-    fn duplicate_names_across_named_and_extras_are_deduplicated() {
-        let def = def_with_scope(ToolScope::Named(vec![
-            "research".into(),
-            "query_memory".into(),
-        ]));
-        let extras: Vec<Box<dyn Tool>> = vec![
-            Box::new(StubTool { name: "research" }), // collides with named
-            Box::new(StubTool { name: "plan" }),
-        ];
-        let set = build_visible_tool_set(&def, &extras).expect("named scope yields Some");
-        assert_eq!(set.len(), 3);
-        assert!(set.contains("research"));
-        assert!(set.contains("query_memory"));
-        assert!(set.contains("plan"));
-    }
-
-    /// `AgentScoping::unscoped` is the safe-fallback constructor used
-    /// when the registry is uninitialised or the target agent isn't
-    /// found. All three fields must default to "no scoping applied"
-    /// so the channel turn runs with the legacy unfiltered behaviour.
-    #[test]
-    fn agent_scoping_unscoped_has_no_filter_or_extras() {
-        let scoping = AgentScoping::unscoped();
-        assert!(scoping.target_agent_id.is_none());
-        assert!(scoping.visible_tool_names.is_none());
-        assert!(scoping.extra_tools.is_empty());
-    }
-}
-
-#[cfg(test)]
-mod approval_surface_gating_tests {
-    use super::channel_has_approval_surface;
-
-    // Sub-issue 2 of #3098: this gate is what decides whether the dispatch
-    // loop sets an `ApprovalChatContext` (→ gate fires for `Prompt`-class
-    // tools) versus the legacy bypass (→ tool calls silently allowed).
-    // Pin the matrix so silently broadening to a new channel can't
-    // accidentally TTL-deny every parked tool call there.
-
-    #[test]
-    fn telegram_has_approval_surface() {
-        assert!(channel_has_approval_surface("telegram"));
-    }
-
-    #[test]
-    fn other_channels_do_not_yet_have_an_approval_surface() {
-        for channel in ["discord", "slack", "imessage", "mattermost", "web", "irc"] {
-            assert!(
-                !channel_has_approval_surface(channel),
-                "channel {channel:?} is not (yet) wired to a per-channel approval surface; \
-                 the dispatch loop must not scope an ApprovalChatContext for it or every \
-                 Prompt-class tool call will park with nobody to answer and TTL-deny"
-            );
-        }
-    }
-
-    #[test]
-    fn unknown_channel_does_not_have_approval_surface() {
-        assert!(!channel_has_approval_surface(""));
-        assert!(!channel_has_approval_surface("Telegram")); // case-sensitive on purpose
-        assert!(!channel_has_approval_surface("telegram-bot"));
-    }
-}
-
-#[cfg(any(test, debug_assertions))]
-pub mod test_support {
-    //! Debug-build seams for raw integration coverage of dispatch helpers.
-
-    use super::*;
-
-    pub fn build_channel_context_block_for_test(msg: &traits::ChannelMessage) -> String {
-        build_channel_context_block(msg)
-    }
-
-    pub fn select_acknowledgment_reaction_for_test(content: &str) -> &'static str {
-        select_acknowledgment_reaction(content)
-    }
-}
+use super::helpers::{
+    build_channel_context_block, log_worker_join_result, select_acknowledgment_reaction,
+    spawn_scoped_typing_task, REPLY_LOG_TRUNCATE_CHARS,
+};
+use super::routing::resolve_target_agent;
 
 /// Whether a channel currently has a registered approval surface — i.e.
 /// a subscriber that turns `ApprovalRequested` events into chat messages
@@ -1135,7 +526,7 @@ pub(crate) async fn process_channel_message(
                 turns.push(ChatMessage::user(&enriched_message));
                 turns.push(ChatMessage::assistant(&response));
                 // Trim to MAX_CHANNEL_HISTORY (keep recent turns)
-                while turns.len() > MAX_CHANNEL_HISTORY {
+                while turns.len() > crate::openhuman::channels::context::MAX_CHANNEL_HISTORY {
                     turns.remove(0);
                 }
             }
@@ -1380,7 +771,3 @@ pub(crate) async fn run_message_dispatch_loop(
         log_worker_join_result(result);
     }
 }
-
-#[cfg(test)]
-#[path = "dispatch_tests.rs"]
-mod tests;
